reformat; added celery tasks
