--- blocking_early_warnings/blocking_early_warnings/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class BlockingEarlyWarningsConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "blocking_early_warnings"
