--- blocking_early_warnings/blocking_early_warnings/settings.py
@@ -1,14 +1,0 @@
-"""
-    File with a few constants to use throughout the app
-"""
-# How many days do we store
-NUMBER_OF_HOURS = 24
-
-# ooni endpoit to request data from
-OONI_ENDPOINT = "https://api.ooni.io/api/v1/measurements"
-
-# Date format
-
-
-DATE_FORMAT = "%Y-%m-%dT%H:%M:%SZ"
-COUNTRY_CODE = "VE"
